+++ docs/CHANGELOG.md
@@ -1,5 +1,22 @@
 # el-form-docs
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - el-form-react@2.0.0
+  - el-form-core@1.0.3
+
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - el-form-react@1.0.2
+  - el-form-core@1.0.2
+
 ## 0.0.1
 
 ### Patch Changes